requirements.txt
astroid==2.11.6
attrs==21.4.0
bleach==5.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
commonmark==0.9.1
dill==0.3.5.1
docutils==0.18.1
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
isort==5.10.1
keyring==23.6.0
lazy-object-proxy==1.7.1
mccabe==0.7.0
packaging==21.3
Pillow==9.2.0
Pillow-PIL==0.1.dev0
pkginfo==1.8.3
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
Pygments==2.12.0
pylint==2.14.4
pyparsing==3.0.9
pytest==7.1.2
pytest-runner==6.0.0
readme-renderer==35.0
requests==2.28.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.4.4
six==1.16.0
tomli==2.0.1
tomlkit==0.11.0
twine==4.0.1
typing-extensions==4.3.0
urllib3==1.26.9
webencodings==0.5.1
wrapt==1.14.1
zipp==3.8.0
